embedded/AppOwnsData/requirements.txt
flask
requests
msal
python-dotenv
authlib
flask-session